+++ android/eir/src/main/java/org/smartregister/fhircore/eir/ui/patient/details/PatientDetailsFragment.kt
@@ -16,11 +16,13 @@
 
 package org.smartregister.fhircore.eir.ui.patient.details
 
+import android.content.DialogInterface
 import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.appcompat.app.AlertDialog
 import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
@@ -40,8 +42,10 @@ import org.hl7.fhir.r4.model.Patient
 import org.smartregister.fhircore.eir.R
 import org.smartregister.fhircore.eir.data.PatientRepository
 import org.smartregister.fhircore.eir.ui.patient.register.PatientItemMapper
-import org.smartregister.fhircore.eir.ui.vaccine.RecordVaccineActivity
 import org.smartregister.fhircore.eir.util.EirConfigClassification
+import org.smartregister.fhircore.eir.ui.adverseevent.AdverseEventQuestionnaireActivity
+import org.smartregister.fhircore.eir.ui.vaccine.RecordVaccineActivity
+import org.smartregister.fhircore.eir.util.ADVERSE_EVENT_FORM
 import org.smartregister.fhircore.eir.util.RECORD_VACCINE_FORM
 import org.smartregister.fhircore.engine.configuration.app.ConfigurableApplication
 import org.smartregister.fhircore.engine.configuration.view.ConfigurableView
@@ -128,13 +132,54 @@ class PatientDetailsFragment private constructor() :
       startActivity(
         Intent(requireContext(), RecordVaccineActivity::class.java)
           .putExtras(
-            QuestionnaireActivity.requiredIntentArgs(
+            QuestionnaireActivity.intentArgs(
               clientIdentifier = patientId,
-              form = RECORD_VACCINE_FORM
+              formName = RECORD_VACCINE_FORM
             )
           )
       )
     }
+
+    reportAdverseEventButton.setOnClickListener {
+      val immunizations = patientDetailsViewModel.patientImmunizations.value as List<Immunization>
+      val immunizationItemWithIds =
+        immunizations.toImmunizationAdverseEventItem(requireContext()).first()
+
+      goToAdverseEventQuestionnaireActivity(immunizationItemWithIds, patientId)
+    }
+  }
+
+  private fun goToAdverseEventQuestionnaireActivity(
+    immunizationAdverseEventItem: ImmunizationAdverseEventItem,
+    patientId: String
+  ) {
+
+    val list = arrayListOf<String>()
+    immunizationAdverseEventItem.dosesWithAdverseEvents.forEach { dose -> list.add(dose.first) }
+
+    AlertDialog.Builder(requireActivity(), R.style.Theme_AppCompat_Light_Dialog_Alert)
+      .apply {
+        setTitle(
+          requireContext()
+            .getString(R.string.choose_dose_adverse_event, immunizationAdverseEventItem.vaccine)
+        )
+        setNegativeButton(R.string.cancel) { dialog: DialogInterface, _ -> dialog.dismiss() }
+        setSingleChoiceItems(list.toTypedArray(), -1) { dialog: DialogInterface, position: Int ->
+          startActivity(
+            Intent(requireContext(), AdverseEventQuestionnaireActivity::class.java)
+              .putExtras(
+                QuestionnaireActivity.intentArgs(
+                  clientIdentifier = patientId,
+                  formName = ADVERSE_EVENT_FORM,
+                  immunizationId = immunizationAdverseEventItem.immunizationIds[position]
+                )
+              )
+          )
+          dialog.dismiss()
+        }
+      }
+      .create()
+      .show()
   }
 
   private fun handlePatientDemographics(patient: Patient) {