+++ android/eir/src/main/java/org/smartregister/fhircore/eir/ui/patient/register/PatientRegisterActivity.kt
@@ -16,10 +16,14 @@
 
 package org.smartregister.fhircore.eir.ui.patient.register
 
+import android.content.Intent
 import android.os.Bundle
+import android.view.View
 import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
+import androidx.lifecycle.Observer
 import org.smartregister.fhircore.eir.R
+import org.smartregister.fhircore.eir.ui.patient.details.PatientDetailsActivity
 import org.smartregister.fhircore.eir.util.EirConfigClassification
 import org.smartregister.fhircore.engine.configuration.view.RegisterViewConfiguration
 import org.smartregister.fhircore.engine.ui.register.BaseRegisterActivity
@@ -53,4 +57,28 @@ class PatientRegisterActivity : BaseRegisterActivity() {
 
   override fun supportedFragments(): Map<String, Fragment> =
     mapOf(Pair(PatientRegisterFragment.TAG, PatientRegisterFragment()))
+
+  override fun onBarcodeResult(barcode: String, view: View) {
+    super.onBarcodeResult(barcode, view)
+
+    isPatientExists(barcode)
+      .observe(
+        this,
+        Observer {
+          if (it.isSuccess) {
+            navigateToDetails(barcode)
+          } else {
+            registerClient(barcode)
+          }
+        }
+      )
+  }
+
+  fun navigateToDetails(patientIdentifier: String) {
+    startActivity(
+      Intent(this, PatientDetailsActivity::class.java).apply {
+        putExtras(PatientDetailsActivity.requiredIntentArgs(patientIdentifier))
+      }
+    )
+  }
 }