+++ android/eir/src/test/java/org/smartregister/fhircore/eir/ui/patient/details/PatientDetailsFragmentTest.kt
@@ -154,7 +154,7 @@ internal class PatientDetailsFragmentTest : FragmentRobolectricTest() {
 
     val patientAgeTextView =
       patientDetailsFragment.view?.findViewById<TextView>(R.id.patientAgeTextView)
-    Assert.assertEquals("Age 0", patientAgeTextView?.text.toString())
+    Assert.assertEquals("Age 0d", patientAgeTextView?.text.toString())
   }
 
   @Test
@@ -258,7 +258,13 @@ internal class PatientDetailsFragmentTest : FragmentRobolectricTest() {
           CodeableConcept(Coding("system", "vaccine_code", "code display")).setText("Astrazeneca")
         protocolApplied =
           listOf(Immunization.ImmunizationProtocolAppliedComponent(PositiveIntType(1)))
-        occurrence = DateTimeType("2021-07-30")
+        occurrence =
+          mockk<DateTimeType>().apply {
+            every { toHumanDisplay() } returns "Sep 16, 2021 6:13:22 PM"
+            every { copy() } returns this
+            every { add(any(), any()) } returns Unit
+            every { after(any()) } returns true
+          }
       }
 
     val immunization2 =
@@ -267,7 +273,13 @@ internal class PatientDetailsFragmentTest : FragmentRobolectricTest() {
           CodeableConcept(Coding("system", "vaccine_code", "code display")).setText("Astrazeneca")
         protocolApplied =
           listOf(Immunization.ImmunizationProtocolAppliedComponent(PositiveIntType(2)))
-        occurrence = DateTimeType("2021-07-30")
+        occurrence =
+          mockk<DateTimeType>().apply {
+            every { toHumanDisplay() } returns "Sep 16, 2021 6:13:22 PM"
+            every { copy() } returns this
+            every { add(any(), any()) } returns Unit
+            every { after(any()) } returns true
+          }
       }
 
     val immunization3 =
@@ -276,7 +288,13 @@ internal class PatientDetailsFragmentTest : FragmentRobolectricTest() {
           CodeableConcept(Coding("system", "vaccine_code", "code display")).setText("Pfizer")
         protocolApplied =
           listOf(Immunization.ImmunizationProtocolAppliedComponent(PositiveIntType(2)))
-        occurrence = DateTimeType("2021-07-30")
+        occurrence =
+          mockk<DateTimeType>().apply {
+            every { toHumanDisplay() } returns "Sep 16, 2021 6:13:22 PM"
+            every { copy() } returns this
+            every { add(any(), any()) } returns Unit
+            every { after(any()) } returns true
+          }
       }
     return listOf(immunization1, immunization2, immunization3)
   }