+++ android/engine/src/main/java/org/smartregister/fhircore/engine/ui/questionnaire/QuestionnaireActivity.kt
@@ -27,14 +27,17 @@ import androidx.fragment.app.commit
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.lifecycleScope
 import ca.uhn.fhir.context.FhirContext
+import com.google.android.fhir.FhirEngineProvider.fhirEngine
 import com.google.android.fhir.datacapture.QuestionnaireFragment
 import com.google.android.fhir.datacapture.QuestionnaireFragment.Companion.BUNDLE_KEY_QUESTIONNAIRE
 import com.google.android.fhir.datacapture.QuestionnaireFragment.Companion.BUNDLE_KEY_QUESTIONNAIRE_RESPONSE
 import com.google.android.fhir.datacapture.validation.QuestionnaireResponseValidator
+import com.google.android.fhir.db.ResourceNotFoundException
 import com.google.android.fhir.logicalId
 import org.hl7.fhir.r4.model.Patient
 import org.hl7.fhir.r4.model.Questionnaire
 import org.hl7.fhir.r4.model.QuestionnaireResponse
+import org.hl7.fhir.r4.model.StringType
 import org.smartregister.fhircore.engine.R
 import org.smartregister.fhircore.engine.configuration.app.ConfigurableApplication
 import org.smartregister.fhircore.engine.ui.base.AlertDialogue.showConfirmAlert
@@ -67,6 +70,10 @@ open class QuestionnaireActivity : BaseMultiLanguageActivity(), View.OnClickList
 
   protected var clientIdentifier: String? = null
 
+  protected var immunizationId: String? = null
+
+  var readOnly: Boolean = false
+
   private val parser = FhirContext.forR4().newJsonParser()
 
   override fun onCreate(savedInstanceState: Bundle?) {
@@ -81,17 +88,19 @@ open class QuestionnaireActivity : BaseMultiLanguageActivity(), View.OnClickList
     val loadProgress = showProgressAlert(this, R.string.loading)
 
     clientIdentifier = intent.getStringExtra(QUESTIONNAIRE_ARG_PATIENT_KEY)
+    immunizationId = intent.getStringExtra(ADVERSE_EVENT_IMMUNIZATION_ITEM_KEY)
 
     lifecycleScope.launchWhenCreated {
-      questionnaireViewModel = createViewModel(application)
+      readOnly = intent.getBooleanExtra(QUESTIONNAIRE_READ_ONLY, false)
+      questionnaireViewModel = createViewModel(application, readOnly)
 
-      val form = intent.getStringExtra(QUESTIONNAIRE_ARG_FORM)!!
+      val formName = intent.getStringExtra(QUESTIONNAIRE_ARG_FORM)!!
       // form is either name of form in asset/form-config or questionnaire-id
       // load from assets and get questionnaire or if not found build it from questionnaire
       questionnaireConfig =
-        kotlin.runCatching { questionnaireViewModel.getQuestionnaireConfig(form) }.getOrElse {
+        kotlin.runCatching { questionnaireViewModel.getQuestionnaireConfig(formName) }.getOrElse {
           // load questionnaire from db and build config
-          questionnaire = questionnaireViewModel.loadQuestionnaire(form)!!
+          questionnaire = questionnaireViewModel.loadQuestionnaire(formName)!!
 
           QuestionnaireConfig(
             appId = (application as ConfigurableApplication).configurationRegistry.appId,
@@ -110,24 +119,49 @@ open class QuestionnaireActivity : BaseMultiLanguageActivity(), View.OnClickList
         title = questionnaireConfig.title
       }
 
+      findViewById<Button>(R.id.btn_save_client_info).apply {
+        setOnClickListener(this@QuestionnaireActivity)
+        if (readOnly) {
+          text = context.getString(R.string.done)
+        }
+      }
+
       // Only add the fragment once, when the activity is first created.
       if (savedInstanceState == null) {
         val fragment =
-          QuestionnaireFragment().apply {
+          FhirCoreQuestionnaireFragment().apply {
             val parsedQuestionnaire = parser.encodeResourceToString(questionnaire)
             arguments =
               when {
-                clientIdentifier == null ->
-                  bundleOf(Pair(BUNDLE_KEY_QUESTIONNAIRE, parsedQuestionnaire))
+                clientIdentifier == null -> {
+                  bundleOf(Pair(BUNDLE_KEY_QUESTIONNAIRE, parsedQuestionnaire)).apply {
+                    val questionnaireResponse = intent.getStringExtra(QUESTIONNAIRE_RESPONSE)
+                    if (readOnly && questionnaireResponse != null) {
+                      putString(BUNDLE_KEY_QUESTIONNAIRE_RESPONSE, questionnaireResponse)
+                    }
+                  }
+                }
                 clientIdentifier != null -> {
-                  val parsedQuestionnaireResponse =
-                    parser.encodeResourceToString(
-                      questionnaireViewModel.generateQuestionnaireResponse(questionnaire!!, intent)
+
+                  try {
+                    fhirEngine.load(Patient::class.java, clientIdentifier!!)
+                    val parsedQuestionnaireResponse =
+                      parser.encodeResourceToString(
+                        questionnaireViewModel.generateQuestionnaireResponse(
+                          questionnaire!!,
+                          intent
+                        )
+                      )
+                    bundleOf(
+                      Pair(BUNDLE_KEY_QUESTIONNAIRE, parser.encodeResourceToString(questionnaire)),
+                      Pair(BUNDLE_KEY_QUESTIONNAIRE_RESPONSE, parsedQuestionnaireResponse)
                     )
-                  bundleOf(
-                    Pair(BUNDLE_KEY_QUESTIONNAIRE, parsedQuestionnaire),
-                    Pair(BUNDLE_KEY_QUESTIONNAIRE_RESPONSE, parsedQuestionnaireResponse)
-                  )
+                  } catch (e: ResourceNotFoundException) {
+                    setBarcode(questionnaire, clientIdentifier!!, true)
+                    bundleOf(
+                      Pair(BUNDLE_KEY_QUESTIONNAIRE, parser.encodeResourceToString(questionnaire))
+                    )
+                  }
                 }
                 else -> bundleOf(Pair(BUNDLE_KEY_QUESTIONNAIRE, parsedQuestionnaire))
               }
@@ -136,25 +170,54 @@ open class QuestionnaireActivity : BaseMultiLanguageActivity(), View.OnClickList
       }
       loadProgress.dismiss()
     }
-
-    findViewById<Button>(R.id.btn_save_client_info).setOnClickListener(this)
   }
 
-  open fun createViewModel(application: Application) =
+  private fun setBarcode(questionnaire: Questionnaire, code: String, readonly: Boolean) {
+    questionnaire.find(QUESTIONNAIRE_ARG_BARCODE_KEY)?.apply {
+      initial =
+        mutableListOf(Questionnaire.QuestionnaireItemInitialComponent().setValue(StringType(code)))
+      readOnly = readonly
+    }
+  }
+
+  private fun Questionnaire.find(linkId: String): Questionnaire.QuestionnaireItemComponent? {
+    return item.find(linkId, null)
+  }
+  private fun List<Questionnaire.QuestionnaireItemComponent>.find(
+    linkId: String,
+    default: Questionnaire.QuestionnaireItemComponent?
+  ): Questionnaire.QuestionnaireItemComponent? {
+    var result = default
+    forEach {
+      if (it.linkId == linkId) {
+        return it
+      } else if (it.item.isNotEmpty()) {
+        result = it.item.find(linkId, result)
+      }
+    }
+
+    return result
+  }
+
+  open fun createViewModel(application: Application, readOnly: Boolean = false) =
     ViewModelProvider(
       this@QuestionnaireActivity,
-      QuestionnaireViewModel(application).createFactory()
+      QuestionnaireViewModel(application, readOnly).createFactory()
     )[QuestionnaireViewModel::class.java]
 
   override fun onClick(view: View) {
     if (view.id == R.id.btn_save_client_info) {
-      showConfirmAlert(
-        context = this,
-        message = R.string.questionnaire_alert_submit_message,
-        title = R.string.questionnaire_alert_submit_title,
-        confirmButtonListener = { handleQuestionnaireSubmit() },
-        confirmButtonText = R.string.questionnaire_alert_submit_button_title
-      )
+      if (readOnly) {
+        finish()
+      } else {
+        showConfirmAlert(
+          context = this,
+          message = R.string.questionnaire_alert_submit_message,
+          title = R.string.questionnaire_alert_submit_title,
+          confirmButtonListener = { handleQuestionnaireSubmit() },
+          confirmButtonText = R.string.questionnaire_alert_submit_button_title
+        )
+      }
     } else {
       showToast(getString(R.string.error_saving_form))
     }
@@ -193,13 +256,12 @@ open class QuestionnaireActivity : BaseMultiLanguageActivity(), View.OnClickList
     questionnaireViewModel.extractionProgress.observe(
       this,
       { result ->
+        saveProcessingAlertDialog.dismiss()
         if (result) {
           finish()
         } else {
           Timber.e("An error occurred during extraction")
         }
-
-        saveProcessingAlertDialog.dismiss()
       }
     )
   }
@@ -229,14 +291,35 @@ open class QuestionnaireActivity : BaseMultiLanguageActivity(), View.OnClickList
     const val QUESTIONNAIRE_POPULATION_RESOURCES = "questionnaire-population-resources"
     const val QUESTIONNAIRE_FRAGMENT_TAG = "questionnaire-fragment-tag"
     const val QUESTIONNAIRE_ARG_PATIENT_KEY = "questionnaire_patient_item_id"
-    const val QUESTIONNAIRE_ARG_FORM = "questionnaire_form"
+    const val ADVERSE_EVENT_IMMUNIZATION_ITEM_KEY = "adverse_event_immunization_item_id"
     const val FORM_CONFIGURATIONS = "configurations/form/form_configurations.json"
+    const val QUESTIONNAIRE_ARG_FORM = "questionnaire-form-name"
+    const val QUESTIONNAIRE_READ_ONLY = "read-only"
+    const val QUESTIONNAIRE_RESPONSE = "questionnaire-response"
+    const val QUESTIONNAIRE_ARG_BARCODE_KEY = "patient-barcode"
+    const val WHO_IDENTIFIER_SYSTEM = "WHO-HCID"
 
-    fun requiredIntentArgs(clientIdentifier: String?, form: String) =
+    fun intentArgs(
+      clientIdentifier: String? = null,
+      formName: String,
+      readOnly: Boolean = false,
+      questionnaireResponse: QuestionnaireResponse? = null,
+      immunizationId: String? = null
+    ) =
       bundleOf(
         Pair(QUESTIONNAIRE_ARG_PATIENT_KEY, clientIdentifier),
-        Pair(QUESTIONNAIRE_ARG_FORM, form),
+        Pair(QUESTIONNAIRE_ARG_FORM, formName),
+        Pair(QUESTIONNAIRE_READ_ONLY, readOnly),
+        Pair(ADVERSE_EVENT_IMMUNIZATION_ITEM_KEY, immunizationId)
       )
+        .apply {
+          if (questionnaireResponse != null) {
+            putString(
+              QUESTIONNAIRE_RESPONSE,
+              FhirContext.forR4().newJsonParser().encodeResourceToString(questionnaireResponse)
+            )
+          }
+        }
   }
 
   override fun onOptionsItemSelected(item: MenuItem): Boolean {
@@ -250,12 +333,16 @@ open class QuestionnaireActivity : BaseMultiLanguageActivity(), View.OnClickList
   }
 
   override fun onBackPressed() {
-    showConfirmAlert(
-      this,
-      R.string.questionnaire_alert_back_pressed_message,
-      R.string.questionnaire_alert_back_pressed_title,
-      { finish() },
-      R.string.questionnaire_alert_back_pressed_button_title
-    )
+    if (readOnly) {
+      finish()
+    } else {
+      showConfirmAlert(
+        this,
+        R.string.questionnaire_alert_back_pressed_message,
+        R.string.questionnaire_alert_back_pressed_title,
+        { finish() },
+        R.string.questionnaire_alert_back_pressed_button_title
+      )
+    }
   }
 }