+++ android/engine/src/main/java/org/smartregister/fhircore/engine/ui/questionnaire/QuestionnaireViewModel.kt
@@ -25,12 +25,14 @@ import androidx.lifecycle.viewModelScope
 import ca.uhn.fhir.context.FhirContext
 import com.google.android.fhir.datacapture.mapping.ResourceMapper
 import com.google.android.fhir.datacapture.targetStructureMap
+import com.google.android.fhir.logicalId
 import java.util.Date
 import java.util.UUID
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import org.hl7.fhir.r4.model.Bundle
+import org.hl7.fhir.r4.model.Identifier
 import org.hl7.fhir.r4.model.Patient
 import org.hl7.fhir.r4.model.Questionnaire
 import org.hl7.fhir.r4.model.QuestionnaireResponse
@@ -48,6 +50,7 @@ import org.smartregister.fhircore.engine.util.helper.TransformSupportServices
 
 open class QuestionnaireViewModel(
   application: Application,
+  private val readOnly: Boolean = false,
 ) : AndroidViewModel(application) {
 
   val extractionProgress = MutableLiveData<Boolean>()
@@ -59,7 +62,12 @@ open class QuestionnaireViewModel(
 
   var structureMapProvider: (suspend (String) -> StructureMap?)? = null
 
-  suspend fun loadQuestionnaire(id: String): Questionnaire? = defaultRepository.loadResource(id)
+  suspend fun loadQuestionnaire(id: String): Questionnaire? =
+    defaultRepository.loadResource<Questionnaire>(id)?.apply {
+      if (readOnly) {
+        changeQuestionsToReadOnly(this.item)
+      }
+    }
 
   suspend fun getQuestionnaireConfig(form: String): QuestionnaireConfig {
     val loadConfig =
@@ -74,7 +82,7 @@ open class QuestionnaireViewModel(
   suspend fun fetchStructureMap(structureMapUrl: String?): StructureMap? {
     var structureMap: StructureMap? = null
     structureMapUrl?.substringAfterLast("/")?.run {
-      structureMap = defaultRepository.loadResource(this)
+      structureMap = loadResource(this) as StructureMap?
     }
     return structureMap
   }
@@ -166,6 +174,10 @@ open class QuestionnaireViewModel(
     return defaultRepository.loadResource(patientId)
   }
 
+  suspend fun loadResource(resourceId: String): Resource? {
+    return defaultRepository.loadResource(resourceId)
+  }
+
   suspend fun loadRelatedPerson(patientId: String): List<RelatedPerson>? {
     return defaultRepository.loadRelatedPersons(patientId)
   }
@@ -183,7 +195,19 @@ open class QuestionnaireViewModel(
     }
 
     intent.getStringExtra(QuestionnaireActivity.QUESTIONNAIRE_ARG_PATIENT_KEY)?.let { patientId ->
-      loadPatient(patientId)?.apply { resourcesList.add(this) }
+      loadPatient(patientId)?.apply {
+        if (identifier.isEmpty()) {
+          identifier =
+            mutableListOf(
+              Identifier().apply {
+                value = logicalId
+                use = Identifier.IdentifierUse.OFFICIAL
+                system = QuestionnaireActivity.WHO_IDENTIFIER_SYSTEM
+              }
+            )
+        }
+        resourcesList.add(this)
+      }
       loadRelatedPerson(patientId)?.forEach { resourcesList.add(it) }
     }
 
@@ -195,5 +219,15 @@ open class QuestionnaireViewModel(
     intent: Intent
   ): QuestionnaireResponse {
     return ResourceMapper.populate(questionnaire, *getPopulationResources(intent))
+  }
+
+  private fun changeQuestionsToReadOnly(items: List<Questionnaire.QuestionnaireItemComponent>) {
+    items.forEach { item ->
+      if (item.type != Questionnaire.QuestionnaireItemType.GROUP) {
+        item.readOnly = true
+      } else {
+        changeQuestionsToReadOnly(item.item)
+      }
+    }
   }
 }