+++ android/engine/src/test/java/org/smartregister/fhircore/engine/ui/questionnaire/QuestionnaireViewModelTest.kt
@@ -17,6 +17,7 @@
 package org.smartregister.fhircore.engine.ui.questionnaire
 
 import android.app.Application
+import android.content.Intent
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.test.core.app.ApplicationProvider
 import com.google.android.fhir.FhirEngine
@@ -37,9 +38,12 @@ import org.hl7.fhir.r4.model.CodeableConcept
 import org.hl7.fhir.r4.model.Coding
 import org.hl7.fhir.r4.model.Expression
 import org.hl7.fhir.r4.model.Extension
+import org.hl7.fhir.r4.model.HumanName
 import org.hl7.fhir.r4.model.Patient
 import org.hl7.fhir.r4.model.Questionnaire
 import org.hl7.fhir.r4.model.QuestionnaireResponse
+import org.hl7.fhir.r4.model.RelatedPerson
+import org.hl7.fhir.r4.model.StringType
 import org.hl7.fhir.r4.model.StructureMap
 import org.junit.Assert
 import org.junit.Before
@@ -83,6 +87,48 @@ class QuestionnaireViewModelTest : RobolectricTest() {
       Questionnaire().apply { id = "12345" }
     val result = runBlocking { questionnaireViewModel.loadQuestionnaire("12345") }
     Assert.assertEquals("12345", result!!.logicalId)
+  }
+
+  @Test
+  fun testLoadQuestionnaireShouldSetQuestionnaireQuestionsToReadOnly() {
+    val questionnaire =
+      Questionnaire().apply {
+        id = "12345"
+        item =
+          listOf(
+            Questionnaire.QuestionnaireItemComponent().apply {
+              type = Questionnaire.QuestionnaireItemType.GROUP
+              linkId = "q1-grp"
+              item =
+                listOf(
+                  Questionnaire.QuestionnaireItemComponent().apply {
+                    type = Questionnaire.QuestionnaireItemType.TEXT
+                    linkId = "q1-name"
+                  }
+                )
+            },
+            Questionnaire.QuestionnaireItemComponent().apply {
+              type = Questionnaire.QuestionnaireItemType.CHOICE
+              linkId = "q2-gender"
+            },
+            Questionnaire.QuestionnaireItemComponent().apply {
+              type = Questionnaire.QuestionnaireItemType.DATE
+              linkId = "q3-date"
+            }
+          )
+      }
+    coEvery { fhirEngine.load(Questionnaire::class.java, "12345") } returns questionnaire
+    questionnaireViewModel = spyk(QuestionnaireViewModel(context, true))
+    ReflectionHelpers.setField(questionnaireViewModel, "defaultRepository", defaultRepo)
+
+    val result = runBlocking { questionnaireViewModel.loadQuestionnaire("12345") }
+
+    Assert.assertTrue(result!!.item[0].item[0].readOnly)
+    Assert.assertEquals("q1-name", result!!.item[0].item[0].linkId)
+    Assert.assertTrue(result.item[1].readOnly)
+    Assert.assertEquals("q2-gender", result.item[1].linkId)
+    Assert.assertTrue(result.item[2].readOnly)
+    Assert.assertEquals("q3-date", result.item[2].linkId)
   }
 
   @Test
@@ -196,5 +242,86 @@ class QuestionnaireViewModelTest : RobolectricTest() {
       questionnaireResponseSlot.captured.subject.reference.replace("Patient/", "")
     )
     Assert.assertEquals("1234567", questionnaireResponseSlot.captured.meta.tagFirstRep.code)
+  }
+
+  @Test
+  fun testLoadPatientShouldReturnPatientResource() {
+    val patient =
+      Patient().apply {
+        name =
+          listOf(
+            HumanName().apply {
+              given = listOf(StringType("John"))
+              family = "Doe"
+            }
+          )
+      }
+    coEvery { fhirEngine.load(Patient::class.java, "1") } returns patient
+
+    runBlocking {
+      val loadedPatient = questionnaireViewModel.loadPatient("1")
+
+      Assert.assertEquals(
+        patient.name.first().given.first().value,
+        loadedPatient?.name?.first()?.given?.first()?.value
+      )
+      Assert.assertEquals(patient.name.first().family, loadedPatient?.name?.first()?.family)
+    }
+  }
+
+  @Test
+  fun testLoadRelatedPersonShouldReturnOnlyOneItemList() {
+    val relatedPerson =
+      RelatedPerson().apply {
+        name =
+          listOf(
+            HumanName().apply {
+              given = listOf(StringType("John"))
+              family = "Doe"
+            }
+          )
+      }
+
+    coEvery { defaultRepo.loadRelatedPersons("1") } returns listOf(relatedPerson)
+
+    runBlocking {
+      val list = questionnaireViewModel.loadRelatedPerson("1")
+      Assert.assertEquals(1, list?.size)
+      val result = list?.get(0)
+      Assert.assertEquals(
+        relatedPerson.name.first().given.first().value,
+        result?.name?.first()?.given?.first()?.value
+      )
+      Assert.assertEquals(relatedPerson.name.first().family, result?.name?.first()?.family)
+    }
+  }
+
+  @Test
+  fun testSaveResourceShouldVerifyResourceSaveMethodCall() {
+    coEvery { defaultRepo.save(any()) } returns Unit
+    questionnaireViewModel.saveResource(mockk())
+    coVerify(exactly = 1) { defaultRepo.save(any()) }
+  }
+
+  @Test
+  fun testGetPopulationResourcesShouldReturnListOfResources() {
+
+    coEvery { fhirEngine.load(Patient::class.java, "2") } returns Patient().apply { id = "2" }
+    coEvery { defaultRepo.loadRelatedPersons("2") } returns
+      listOf(RelatedPerson().apply { id = "3" })
+
+    val intent = Intent()
+    intent.putStringArrayListExtra(
+      QuestionnaireActivity.QUESTIONNAIRE_POPULATION_RESOURCES,
+      arrayListOf(
+        "{\"resourceType\":\"Patient\",\"id\":\"1\",\"text\":{\"status\":\"generated\",\"div\":\"\"}}"
+      )
+    )
+    intent.putExtra(QuestionnaireActivity.QUESTIONNAIRE_ARG_PATIENT_KEY, "2")
+
+    runBlocking {
+      val resourceList = questionnaireViewModel.getPopulationResources(intent)
+      Assert.assertEquals(3, resourceList.size)
+    }
   }
 }