+++ CHANGELOG.md
@@ -8,7 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased] - 2021-xx-xx
 
 ### Added
+- EIR | Add Adverse Events feature to log any reactions after the first dose of vaccine using Structure map
 - Quest | Add the ability to search for patients by id
+- Quest | Add ability to view previous QuestionnaireResponse
 
 ### Fixed
 -
@@ -40,6 +42,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Engine | Block questionnaire submit if validation errors exist
 - Quest | Bottom Navigation updated with new structure
 - Quest | User actions, profile screen implemented
+- Engine | print patient age as #y #m or #m #d
 - ANC | Navigation structure changes implemented
 - ANC | Bottom sheet for switching registers added
 - ANC | User actions, profile screen implemented
@@ -52,6 +55,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Authentication toekn expiry issue
 - Fhir Resource Converter issue after resource update
 - Inteceptor handling for missing account
+- Engine: Fixes Structure map based resource extraction not working | Remove the menu icon on the login screen
 - ANC | Fix crashes due to invalid data
 - Engine | Remove the menu icon on the login screen
 - Engine: Fixes Structure map based resource extraction not working